src/website/requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
requests
pydub
pysubs2